修复JPS算法中线性插值的问题； 修改写入的规划时间的格式，从long->double； 修复地图长宽读取顺序相反的问题 修复终点选举逻辑的问题（问题是，跳过了不该跳过的点，导致有些小车在完成自己的区域后就来回踱步的问题

+++ src/main/java/ncepusa/distributedcars/navigator/data_structures/GridNode.java
@@ -99,11 +99,11 @@ public class GridNode implements Comparable<GridNode> {
     public boolean equals(Object o) {
         if (o == null || getClass() != o.getClass()) return false;
         GridNode gridNode = (GridNode) o;
-        return position.getX() == gridNode.getXY().getX() && position.getY() == gridNode.getXY().getY();
+        return (int)position.getX() == (int)gridNode.getXY().getX() && (int)position.getY() == (int)gridNode.getXY().getY();
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(position.getX(), position.getY());
+        return Objects.hash((int)position.getX(), (int)position.getY());
     }
 }
